CAS-587: Refactor common components
