+++ frontend/packages/client/src/App.sass
@@ -172,8 +172,6 @@ code
 	background-color: $orange
 .has-background-blue
 	background-color: $light-blue
-.has-background-green
-	background-color: $success
 .has-background-grey
 	background-color: $grey
 .has-background-blue
@@ -184,8 +182,6 @@ code
 	background-color: hsl(0deg, 0%, 86%, 0.3) !important
 .has-background-yellow
 	background-color: $yellow
-.has-background-green
-	background-color: $success
 	&.button
 		border: none !important
 		&:hover
@@ -193,6 +189,8 @@ code
 		&.vote-button:hover, &.has-button-border-hover:hover
 			color: black
 			border: 1px solid rgba(0, 0, 0, 0.4) !important
+.has-background-green
+	background-color: $success
 .button.is-disabled
 	pointer-events: none
 	opacity: 0.5
